refactor(web-ui): add new hook useRegisterConfirm Extract logical from component to hook
